Reverted LoginController loginSuccessful method, added setting Main.controller functionality to initialize method in PanelController Fixed potential null pointer in DiscordCommunication.checkToken Added spinner functionality OBSTHObject now retrieves current Y pos instead of forcing user to enter one Added OBSTHObject.refresh()

+++ src/test/java/OBSObjectTest.java
@@ -7,7 +7,7 @@ public class OBSObjectTest {
                 res -> {
                     System.out.println(res.toString());
                     System.out.println(res.getSceneItemId());
-                    OBSTHObject test = new OBSTHObject("test", "Image", (float)873);
+                    OBSTHObject test = new OBSTHObject("test", "Image", 873);
                     test.sceneItemId = 1;
                     try {
                         while (true) {
